Add setting to enable/disable the screenshot feature, #1066
