+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add setting fields to customize the metabox title and checkbox label (Issue #227).
 - Add method `disableExpiration` to the class `PublishPress\Future\Modules\Expirator\Models\ExpirablePostModel`.
 - Add `*` to indicate required fields in the workflow editor (Issue #974).
+- Add new setting to disable/enable the screenshot feature in the workflow editor (Issue #1066).
 
 ### Changed
 
 - Move notification settings to a specific tab (Issue #190).
+- Disable the workflow screenshot feature by default (Issue #1066).
 
 ### Fixed
