+++ assets/js/workflowEditor.js
@@ -16480,8 +16480,6 @@ var LayoutHeader = function LayoutHeader() {
         isInserterOpened: select(_editor_store__WEBPACK_IMPORTED_MODULE_5__.store).isFeatureActive(_constants__WEBPACK_IMPORTED_MODULE_7__.FEATURE_INSERTER),
         showIconLabels: select(_editor_store__WEBPACK_IMPORTED_MODULE_5__.store).isFeatureActive(_constants__WEBPACK_IMPORTED_MODULE_7__.FEATURE_SHOW_ICON_LABELS),
         isLoadingWorkflow: select(_workflow_store__WEBPACK_IMPORTED_MODULE_6__.store).isLoadingWorkflow(),
-        isEditedWorkflowDirty: select(_workflow_store__WEBPACK_IMPORTED_MODULE_6__.store).isEditedWorkflowDirty(),
-        takeScreenshot: select(_workflow_store__WEBPACK_IMPORTED_MODULE_6__.store).takeScreenshot,
         isPro: select(_editor_store__WEBPACK_IMPORTED_MODULE_5__.store).isPro()
       };
     }),
@@ -16490,8 +16488,6 @@ var LayoutHeader = function LayoutHeader() {
     hasReducedUI = _useSelect.hasReducedUI,
     showIconLabels = _useSelect.showIconLabels,
     isLoadingWorkflow = _useSelect.isLoadingWorkflow,
-    isEditedWorkflowDirty = _useSelect.isEditedWorkflowDirty,
-    takeScreenshot = _useSelect.takeScreenshot,
     isPro = _useSelect.isPro;
   var _useDispatch = (0,_wordpress_data__WEBPACK_IMPORTED_MODULE_0__.useDispatch)(_editor_store__WEBPACK_IMPORTED_MODULE_5__.store),
     enableFeature = _useDispatch.enableFeature,
@@ -21226,6 +21222,8 @@ function WorkflowPublishButton(_ref) {
     saveAsCurrentStatus = _useDispatch.saveAsCurrentStatus,
     publishWorkflow = _useDispatch.publishWorkflow;
   var wasSaving = (0,_wordpress_compose__WEBPACK_IMPORTED_MODULE_7__["default"])(isSaving);
+  var _futureWorkflowEditor = futureWorkflowEditor,
+    enableWorkflowScreenshot = _futureWorkflowEditor.enableWorkflowScreenshot;
   (0,_wordpress_element__WEBPACK_IMPORTED_MODULE_3__.useEffect)(function () {
     var timeoutId;
     if (wasSaving && !isSaving) {
@@ -21259,17 +21257,25 @@ function WorkflowPublishButton(_ref) {
     if (isDisabled) {
       return;
     }
-    takeScreenshot().then(function (dataUrl) {
+    if (enableWorkflowScreenshot) {
+      takeScreenshot().then(function (dataUrl) {
+        if (isPublished) {
+          saveAsCurrentStatus({
+            screenshot: dataUrl
+          });
+        } else {
+          publishWorkflow({
+            screenshot: dataUrl
+          });
+        }
+      });
+    } else {
       if (isPublished) {
-        saveAsCurrentStatus({
-          screenshot: dataUrl
-        });
+        saveAsCurrentStatus();
       } else {
-        publishWorkflow({
-          screenshot: dataUrl
-        });
+        publishWorkflow();
       }
-    });
+    }
   };
 
   // Use common Button instance for all saved states so that focus is not
@@ -21390,6 +21396,8 @@ function WorkflowSaveDraftButton(_ref) {
   var _useDispatch = (0,_wordpress_data__WEBPACK_IMPORTED_MODULE_2__.useDispatch)(_workflow_store__WEBPACK_IMPORTED_MODULE_5__.store),
     saveAsDraft = _useDispatch.saveAsDraft;
   var wasSaving = (0,_wordpress_compose__WEBPACK_IMPORTED_MODULE_7__["default"])(isSaving);
+  var _futureWorkflowEditor = futureWorkflowEditor,
+    enableWorkflowScreenshot = _futureWorkflowEditor.enableWorkflowScreenshot;
   (0,_wordpress_element__WEBPACK_IMPORTED_MODULE_3__.useEffect)(function () {
     var timeoutId;
     if (wasSaving && !isSaving) {
@@ -21425,11 +21433,15 @@ function WorkflowSaveDraftButton(_ref) {
     text = shortLabel;
   }
   var onClick = function onClick() {
-    takeScreenshot().then(function (screenshot) {
-      saveAsDraft({
-        screenshot: screenshot
+    if (enableWorkflowScreenshot) {
+      takeScreenshot().then(function (screenshot) {
+        saveAsDraft({
+          screenshot: screenshot
+        });
       });
-    });
+    } else {
+      saveAsDraft();
+    }
   };
 
   // Use common Button instance for all saved states so that focus is not
@@ -21604,8 +21616,9 @@ function setupEditor(workflowId) {
   }, _marked, null, [[6, 14], [20, 28]]);
 }
 ;
-function saveAsDraft(_ref) {
-  var screenshot = _ref.screenshot;
+function saveAsDraft() {
+  var _ref = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {},
+    screenshot = _ref.screenshot;
   return /*#__PURE__*/_regeneratorRuntime().mark(function _callee() {
     var wasNewWorkflow, editedWorkflow, newWorkflow;
     return _regeneratorRuntime().wrap(function _callee$(_context2) {
@@ -21673,8 +21686,9 @@ function saveAsDraft(_ref) {
     }, _callee, null, [[2, 21]]);
   })();
 }
-function saveAsCurrentStatus(_ref2) {
-  var screenshot = _ref2.screenshot;
+function saveAsCurrentStatus() {
+  var _ref2 = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {},
+    screenshot = _ref2.screenshot;
   return /*#__PURE__*/_regeneratorRuntime().mark(function _callee2() {
     var editedWorkflow, newWorkflow;
     return _regeneratorRuntime().wrap(function _callee2$(_context3) {
@@ -21731,8 +21745,9 @@ function saveAsCurrentStatus(_ref2) {
     }, _callee2, null, [[2, 15]]);
   })();
 }
-function publishWorkflow(_ref3) {
-  var screenshot = _ref3.screenshot;
+function publishWorkflow() {
+  var _ref3 = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {},
+    screenshot = _ref3.screenshot;
   return /*#__PURE__*/_regeneratorRuntime().mark(function _callee3() {
     var wasNewWorkflow, editedWorkflow, newWorkflow;
     return _regeneratorRuntime().wrap(function _callee3$(_context4) {
@@ -21800,8 +21815,9 @@ function publishWorkflow(_ref3) {
     }, _callee3, null, [[2, 21]]);
   })();
 }
-function switchToDraft(_ref4) {
-  var screenshot = _ref4.screenshot;
+function switchToDraft() {
+  var _ref4 = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {},
+    screenshot = _ref4.screenshot;
   return /*#__PURE__*/_regeneratorRuntime().mark(function _callee4() {
     var wasNewWorkflow, editedWorkflow, newWorkflow;
     return _regeneratorRuntime().wrap(function _callee4$(_context5) {
@@ -22973,15 +22989,15 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export */   isWorkflowFlowEmpty: () => (/* binding */ isWorkflowFlowEmpty),
 /* harmony export */   takeScreenshot: () => (/* binding */ takeScreenshot)
 /* harmony export */ });
-/* harmony import */ var html_to_image__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! html-to-image */ "./node_modules/html-to-image/es/index.js");
+function _regeneratorRuntime() { "use strict"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return e; }; var t, e = {}, r = Object.prototype, n = r.hasOwnProperty, o = Object.defineProperty || function (t, e, r) { t[e] = r.value; }, i = "function" == typeof Symbol ? Symbol : {}, a = i.iterator || "@@iterator", c = i.asyncIterator || "@@asyncIterator", u = i.toStringTag || "@@toStringTag"; function define(t, e, r) { return Object.defineProperty(t, e, { value: r, enumerable: !0, configurable: !0, writable: !0 }), t[e]; } try { define({}, ""); } catch (t) { define = function define(t, e, r) { return t[e] = r; }; } function wrap(t, e, r, n) { var i = e && e.prototype instanceof Generator ? e : Generator, a = Object.create(i.prototype), c = new Context(n || []); return o(a, "_invoke", { value: makeInvokeMethod(t, r, c) }), a; } function tryCatch(t, e, r) { try { return { type: "normal", arg: t.call(e, r) }; } catch (t) { return { type: "throw", arg: t }; } } e.wrap = wrap; var h = "suspendedStart", l = "suspendedYield", f = "executing", s = "completed", y = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var p = {}; define(p, a, function () { return this; }); var d = Object.getPrototypeOf, v = d && d(d(values([]))); v && v !== r && n.call(v, a) && (p = v); var g = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(p); function defineIteratorMethods(t) { ["next", "throw", "return"].forEach(function (e) { define(t, e, function (t) { return this._invoke(e, t); }); }); } function AsyncIterator(t, e) { function invoke(r, o, i, a) { var c = tryCatch(t[r], t, o); if ("throw" !== c.type) { var u = c.arg, h = u.value; return h && "object" == _typeof(h) && n.call(h, "__await") ? e.resolve(h.__await).then(function (t) { invoke("next", t, i, a); }, function (t) { invoke("throw", t, i, a); }) : e.resolve(h).then(function (t) { u.value = t, i(u); }, function (t) { return invoke("throw", t, i, a); }); } a(c.arg); } var r; o(this, "_invoke", { value: function value(t, n) { function callInvokeWithMethodAndArg() { return new e(function (e, r) { invoke(t, n, e, r); }); } return r = r ? r.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); } }); } function makeInvokeMethod(e, r, n) { var o = h; return function (i, a) { if (o === f) throw Error("Generator is already running"); if (o === s) { if ("throw" === i) throw a; return { value: t, done: !0 }; } for (n.method = i, n.arg = a;;) { var c = n.delegate; if (c) { var u = maybeInvokeDelegate(c, n); if (u) { if (u === y) continue; return u; } } if ("next" === n.method) n.sent = n._sent = n.arg;else if ("throw" === n.method) { if (o === h) throw o = s, n.arg; n.dispatchException(n.arg); } else "return" === n.method && n.abrupt("return", n.arg); o = f; var p = tryCatch(e, r, n); if ("normal" === p.type) { if (o = n.done ? s : l, p.arg === y) continue; return { value: p.arg, done: n.done }; } "throw" === p.type && (o = s, n.method = "throw", n.arg = p.arg); } }; } function maybeInvokeDelegate(e, r) { var n = r.method, o = e.iterator[n]; if (o === t) return r.delegate = null, "throw" === n && e.iterator.return && (r.method = "return", r.arg = t, maybeInvokeDelegate(e, r), "throw" === r.method) || "return" !== n && (r.method = "throw", r.arg = new TypeError("The iterator does not provide a '" + n + "' method")), y; var i = tryCatch(o, e.iterator, r.arg); if ("throw" === i.type) return r.method = "throw", r.arg = i.arg, r.delegate = null, y; var a = i.arg; return a ? a.done ? (r[e.resultName] = a.value, r.next = e.nextLoc, "return" !== r.method && (r.method = "next", r.arg = t), r.delegate = null, y) : a : (r.method = "throw", r.arg = new TypeError("iterator result is not an object"), r.delegate = null, y); } function pushTryEntry(t) { var e = { tryLoc: t[0] }; 1 in t && (e.catchLoc = t[1]), 2 in t && (e.finallyLoc = t[2], e.afterLoc = t[3]), this.tryEntries.push(e); } function resetTryEntry(t) { var e = t.completion || {}; e.type = "normal", delete e.arg, t.completion = e; } function Context(t) { this.tryEntries = [{ tryLoc: "root" }], t.forEach(pushTryEntry, this), this.reset(!0); } function values(e) { if (e || "" === e) { var r = e[a]; if (r) return r.call(e); if ("function" == typeof e.next) return e; if (!isNaN(e.length)) { var o = -1, i = function next() { for (; ++o < e.length;) if (n.call(e, o)) return next.value = e[o], next.done = !1, next; return next.value = t, next.done = !0, next; }; return i.next = i; } } throw new TypeError(_typeof(e) + " is not iterable"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, o(g, "constructor", { value: GeneratorFunctionPrototype, configurable: !0 }), o(GeneratorFunctionPrototype, "constructor", { value: GeneratorFunction, configurable: !0 }), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, u, "GeneratorFunction"), e.isGeneratorFunction = function (t) { var e = "function" == typeof t && t.constructor; return !!e && (e === GeneratorFunction || "GeneratorFunction" === (e.displayName || e.name)); }, e.mark = function (t) { return Object.setPrototypeOf ? Object.setPrototypeOf(t, GeneratorFunctionPrototype) : (t.__proto__ = GeneratorFunctionPrototype, define(t, u, "GeneratorFunction")), t.prototype = Object.create(g), t; }, e.awrap = function (t) { return { __await: t }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, c, function () { return this; }), e.AsyncIterator = AsyncIterator, e.async = function (t, r, n, o, i) { void 0 === i && (i = Promise); var a = new AsyncIterator(wrap(t, r, n, o), i); return e.isGeneratorFunction(r) ? a : a.next().then(function (t) { return t.done ? t.value : a.next(); }); }, defineIteratorMethods(g), define(g, u, "Generator"), define(g, a, function () { return this; }), define(g, "toString", function () { return "[object Generator]"; }), e.keys = function (t) { var e = Object(t), r = []; for (var n in e) r.push(n); return r.reverse(), function next() { for (; r.length;) { var t = r.pop(); if (t in e) return next.value = t, next.done = !1, next; } return next.done = !0, next; }; }, e.values = values, Context.prototype = { constructor: Context, reset: function reset(e) { if (this.prev = 0, this.next = 0, this.sent = this._sent = t, this.done = !1, this.delegate = null, this.method = "next", this.arg = t, this.tryEntries.forEach(resetTryEntry), !e) for (var r in this) "t" === r.charAt(0) && n.call(this, r) && !isNaN(+r.slice(1)) && (this[r] = t); }, stop: function stop() { this.done = !0; var t = this.tryEntries[0].completion; if ("throw" === t.type) throw t.arg; return this.rval; }, dispatchException: function dispatchException(e) { if (this.done) throw e; var r = this; function handle(n, o) { return a.type = "throw", a.arg = e, r.next = n, o && (r.method = "next", r.arg = t), !!o; } for (var o = this.tryEntries.length - 1; o >= 0; --o) { var i = this.tryEntries[o], a = i.completion; if ("root" === i.tryLoc) return handle("end"); if (i.tryLoc <= this.prev) { var c = n.call(i, "catchLoc"), u = n.call(i, "finallyLoc"); if (c && u) { if (this.prev < i.catchLoc) return handle(i.catchLoc, !0); if (this.prev < i.finallyLoc) return handle(i.finallyLoc); } else if (c) { if (this.prev < i.catchLoc) return handle(i.catchLoc, !0); } else { if (!u) throw Error("try statement without catch or finally"); if (this.prev < i.finallyLoc) return handle(i.finallyLoc); } } } }, abrupt: function abrupt(t, e) { for (var r = this.tryEntries.length - 1; r >= 0; --r) { var o = this.tryEntries[r]; if (o.tryLoc <= this.prev && n.call(o, "finallyLoc") && this.prev < o.finallyLoc) { var i = o; break; } } i && ("break" === t || "continue" === t) && i.tryLoc <= e && e <= i.finallyLoc && (i = null); var a = i ? i.completion : {}; return a.type = t, a.arg = e, i ? (this.method = "next", this.next = i.finallyLoc, y) : this.complete(a); }, complete: function complete(t, e) { if ("throw" === t.type) throw t.arg; return "break" === t.type || "continue" === t.type ? this.next = t.arg : "return" === t.type ? (this.rval = this.arg = t.arg, this.method = "return", this.next = "end") : "normal" === t.type && e && (this.next = e), y; }, finish: function finish(t) { for (var e = this.tryEntries.length - 1; e >= 0; --e) { var r = this.tryEntries[e]; if (r.finallyLoc === t) return this.complete(r.completion, r.afterLoc), resetTryEntry(r), y; } }, catch: function _catch(t) { for (var e = this.tryEntries.length - 1; e >= 0; --e) { var r = this.tryEntries[e]; if (r.tryLoc === t) { var n = r.completion; if ("throw" === n.type) { var o = n.arg; resetTryEntry(r); } return o; } } throw Error("illegal catch attempt"); }, delegateYield: function delegateYield(e, r, n) { return this.delegate = { iterator: values(e), resultName: r, nextLoc: n }, "next" === this.method && (this.arg = t), y; } }, e; }
 function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
+function asyncGeneratorStep(n, t, e, r, o, a, c) { try { var i = n[a](c), u = i.value; } catch (n) { return void e(n); } i.done ? t(u) : Promise.resolve(u).then(r, o); }
+function _asyncToGenerator(n) { return function () { var t = this, e = arguments; return new Promise(function (r, o) { var a = n.apply(t, e); function _next(n) { asyncGeneratorStep(a, r, o, _next, _throw, "next", n); } function _throw(n) { asyncGeneratorStep(a, r, o, _next, _throw, "throw", n); } _next(void 0); }); }; }
 function ownKeys(e, r) { var t = Object.keys(e); if (Object.getOwnPropertySymbols) { var o = Object.getOwnPropertySymbols(e); r && (o = o.filter(function (r) { return Object.getOwnPropertyDescriptor(e, r).enumerable; })), t.push.apply(t, o); } return t; }
 function _objectSpread(e) { for (var r = 1; r < arguments.length; r++) { var t = null != arguments[r] ? arguments[r] : {}; r % 2 ? ownKeys(Object(t), !0).forEach(function (r) { _defineProperty(e, r, t[r]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(t)) : ownKeys(Object(t)).forEach(function (r) { Object.defineProperty(e, r, Object.getOwnPropertyDescriptor(t, r)); }); } return e; }
 function _defineProperty(e, r, t) { return (r = _toPropertyKey(r)) in e ? Object.defineProperty(e, r, { value: t, enumerable: !0, configurable: !0, writable: !0 }) : e[r] = t, e; }
 function _toPropertyKey(t) { var i = _toPrimitive(t, "string"); return "symbol" == _typeof(i) ? i : i + ""; }
 function _toPrimitive(t, r) { if ("object" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || "default"); if ("object" != _typeof(i)) return i; throw new TypeError("@@toPrimitive must return a primitive value."); } return ("string" === r ? String : Number)(t); }
-// import { createRegistrySelector } from '@wordpress/data';
-
 var getPostType = function getPostType(state) {
   return state.postType;
 };
@@ -23092,26 +23108,51 @@ var getDataTypeByName = function getDataTypeByName(state, name) {
   });
 };
 function takeScreenshot() {
-  return new Promise(function (resolve, reject) {
-    try {
-      (0,html_to_image__WEBPACK_IMPORTED_MODULE_0__.toPng)(document.querySelector('.react-flow'), {
-        filter: function filter(node) {
-          var _node$classList, _node$classList2, _node$classList3;
-          // We don't want to add the minimap and the controls to the image
-          if (node !== null && node !== void 0 && (_node$classList = node.classList) !== null && _node$classList !== void 0 && _node$classList.contains('react-flow__minimap') || node !== null && node !== void 0 && (_node$classList2 = node.classList) !== null && _node$classList2 !== void 0 && _node$classList2.contains('react-flow__controls') || node !== null && node !== void 0 && (_node$classList3 = node.classList) !== null && _node$classList3 !== void 0 && _node$classList3.contains('pwe-node-edit-button')) {
-            return false;
+  return _takeScreenshot.apply(this, arguments);
+}
+function _takeScreenshot() {
+  _takeScreenshot = _asyncToGenerator(/*#__PURE__*/_regeneratorRuntime().mark(function _callee() {
+    var _futureWorkflowEditor, enableWorkflowScreenshot, _yield$import, toPng;
+    return _regeneratorRuntime().wrap(function _callee$(_context) {
+      while (1) switch (_context.prev = _context.next) {
+        case 0:
+          _futureWorkflowEditor = futureWorkflowEditor, enableWorkflowScreenshot = _futureWorkflowEditor.enableWorkflowScreenshot;
+          if (enableWorkflowScreenshot) {
+            _context.next = 3;
+            break;
           }
-          return true;
-        }
-      }).then(function (dataUrl) {
-        resolve(dataUrl); // Resolve with the image dataUrl
-      }).catch(function (error) {
-        reject(error); // Reject with the error
-      });
-    } catch (error) {
-      reject(error); // Reject with the error
-    }
-  });
+          return _context.abrupt("return", null);
+        case 3:
+          _context.prev = 3;
+          _context.next = 6;
+          return __webpack_require__.e(/*! import() */ "vendors-node_modules_html-to-image_es_index_js").then(__webpack_require__.bind(__webpack_require__, /*! html-to-image */ "./node_modules/html-to-image/es/index.js"));
+        case 6:
+          _yield$import = _context.sent;
+          toPng = _yield$import.toPng;
+          _context.next = 10;
+          return toPng(document.querySelector('.react-flow'), {
+            filter: function filter(node) {
+              var _node$classList, _node$classList2, _node$classList3;
+              if (node !== null && node !== void 0 && (_node$classList = node.classList) !== null && _node$classList !== void 0 && _node$classList.contains('react-flow__minimap') || node !== null && node !== void 0 && (_node$classList2 = node.classList) !== null && _node$classList2 !== void 0 && _node$classList2.contains('react-flow__controls') || node !== null && node !== void 0 && (_node$classList3 = node.classList) !== null && _node$classList3 !== void 0 && _node$classList3.contains('pwe-node-edit-button')) {
+                return false;
+              }
+              return true;
+            }
+          });
+        case 10:
+          return _context.abrupt("return", _context.sent);
+        case 13:
+          _context.prev = 13;
+          _context.t0 = _context["catch"](3);
+          console.error('Error loading html-to-image:', _context.t0);
+          return _context.abrupt("return", null);
+        case 17:
+        case "end":
+          return _context.stop();
+      }
+    }, _callee, null, [[3, 13]]);
+  }));
+  return _takeScreenshot.apply(this, arguments);
 }
 function getGlobalVariables(state) {
   return state.globalVariables;
@@ -23274,6 +23315,8 @@ function WorkflowSwitchToDraftButton() {
   if (!isPublishedWorkflow) {
     return null;
   }
+  var _futureWorkflowEditor = futureWorkflowEditor,
+    enableWorkflowScreenshot = _futureWorkflowEditor.enableWorkflowScreenshot;
   var onSwitch = function onSwitch() {
     var alertMessage;
     if (isPublishedWorkflow) {
@@ -23283,11 +23326,15 @@ function WorkflowSwitchToDraftButton() {
       return;
     }
     if (window.confirm(alertMessage)) {
-      takeScreenshot().then(function (dataUrl) {
-        switchToDraft({
-          screenshot: dataUrl
+      if (enableWorkflowScreenshot) {
+        takeScreenshot().then(function (dataUrl) {
+          switchToDraft({
+            screenshot: dataUrl
+          });
         });
-      });
+      } else {
+        switchToDraft();
+      }
     }
   };
   return /*#__PURE__*/React.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_0__.Button, {
@@ -25694,1081 +25741,6 @@ module.exports = function (item) {
   }
   return [content].join("\n");
 };
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/apply-style.js":
-/*!******************************************************!*\
-  !*** ./node_modules/html-to-image/es/apply-style.js ***!
-  \******************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   applyStyle: () => (/* binding */ applyStyle)
-/* harmony export */ });
-function applyStyle(node, options) {
-    const { style } = node;
-    if (options.backgroundColor) {
-        style.backgroundColor = options.backgroundColor;
-    }
-    if (options.width) {
-        style.width = `${options.width}px`;
-    }
-    if (options.height) {
-        style.height = `${options.height}px`;
-    }
-    const manual = options.style;
-    if (manual != null) {
-        Object.keys(manual).forEach((key) => {
-            style[key] = manual[key];
-        });
-    }
-    return node;
-}
-//# sourceMappingURL=apply-style.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/clone-node.js":
-/*!*****************************************************!*\
-  !*** ./node_modules/html-to-image/es/clone-node.js ***!
-  \*****************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   cloneNode: () => (/* binding */ cloneNode)
-/* harmony export */ });
-/* harmony import */ var _clone_pseudos__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./clone-pseudos */ "./node_modules/html-to-image/es/clone-pseudos.js");
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-/* harmony import */ var _mimes__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./mimes */ "./node_modules/html-to-image/es/mimes.js");
-/* harmony import */ var _dataurl__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./dataurl */ "./node_modules/html-to-image/es/dataurl.js");
-
-
-
-
-async function cloneCanvasElement(canvas) {
-    const dataURL = canvas.toDataURL();
-    if (dataURL === 'data:,') {
-        return canvas.cloneNode(false);
-    }
-    return (0,_util__WEBPACK_IMPORTED_MODULE_1__.createImage)(dataURL);
-}
-async function cloneVideoElement(video, options) {
-    if (video.currentSrc) {
-        const canvas = document.createElement('canvas');
-        const ctx = canvas.getContext('2d');
-        canvas.width = video.clientWidth;
-        canvas.height = video.clientHeight;
-        ctx === null || ctx === void 0 ? void 0 : ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
-        const dataURL = canvas.toDataURL();
-        return (0,_util__WEBPACK_IMPORTED_MODULE_1__.createImage)(dataURL);
-    }
-    const poster = video.poster;
-    const contentType = (0,_mimes__WEBPACK_IMPORTED_MODULE_2__.getMimeType)(poster);
-    const dataURL = await (0,_dataurl__WEBPACK_IMPORTED_MODULE_3__.resourceToDataURL)(poster, contentType, options);
-    return (0,_util__WEBPACK_IMPORTED_MODULE_1__.createImage)(dataURL);
-}
-async function cloneIFrameElement(iframe) {
-    var _a;
-    try {
-        if ((_a = iframe === null || iframe === void 0 ? void 0 : iframe.contentDocument) === null || _a === void 0 ? void 0 : _a.body) {
-            return (await cloneNode(iframe.contentDocument.body, {}, true));
-        }
-    }
-    catch (_b) {
-        // Failed to clone iframe
-    }
-    return iframe.cloneNode(false);
-}
-async function cloneSingleNode(node, options) {
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(node, HTMLCanvasElement)) {
-        return cloneCanvasElement(node);
-    }
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(node, HTMLVideoElement)) {
-        return cloneVideoElement(node, options);
-    }
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(node, HTMLIFrameElement)) {
-        return cloneIFrameElement(node);
-    }
-    return node.cloneNode(false);
-}
-const isSlotElement = (node) => node.tagName != null && node.tagName.toUpperCase() === 'SLOT';
-async function cloneChildren(nativeNode, clonedNode, options) {
-    var _a, _b;
-    let children = [];
-    if (isSlotElement(nativeNode) && nativeNode.assignedNodes) {
-        children = (0,_util__WEBPACK_IMPORTED_MODULE_1__.toArray)(nativeNode.assignedNodes());
-    }
-    else if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLIFrameElement) &&
-        ((_a = nativeNode.contentDocument) === null || _a === void 0 ? void 0 : _a.body)) {
-        children = (0,_util__WEBPACK_IMPORTED_MODULE_1__.toArray)(nativeNode.contentDocument.body.childNodes);
-    }
-    else {
-        children = (0,_util__WEBPACK_IMPORTED_MODULE_1__.toArray)(((_b = nativeNode.shadowRoot) !== null && _b !== void 0 ? _b : nativeNode).childNodes);
-    }
-    if (children.length === 0 ||
-        (0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLVideoElement)) {
-        return clonedNode;
-    }
-    await children.reduce((deferred, child) => deferred
-        .then(() => cloneNode(child, options))
-        .then((clonedChild) => {
-        if (clonedChild) {
-            clonedNode.appendChild(clonedChild);
-        }
-    }), Promise.resolve());
-    return clonedNode;
-}
-function cloneCSSStyle(nativeNode, clonedNode) {
-    const targetStyle = clonedNode.style;
-    if (!targetStyle) {
-        return;
-    }
-    const sourceStyle = window.getComputedStyle(nativeNode);
-    if (sourceStyle.cssText) {
-        targetStyle.cssText = sourceStyle.cssText;
-        targetStyle.transformOrigin = sourceStyle.transformOrigin;
-    }
-    else {
-        (0,_util__WEBPACK_IMPORTED_MODULE_1__.toArray)(sourceStyle).forEach((name) => {
-            let value = sourceStyle.getPropertyValue(name);
-            if (name === 'font-size' && value.endsWith('px')) {
-                const reducedFont = Math.floor(parseFloat(value.substring(0, value.length - 2))) - 0.1;
-                value = `${reducedFont}px`;
-            }
-            if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLIFrameElement) &&
-                name === 'display' &&
-                value === 'inline') {
-                value = 'block';
-            }
-            if (name === 'd' && clonedNode.getAttribute('d')) {
-                value = `path(${clonedNode.getAttribute('d')})`;
-            }
-            targetStyle.setProperty(name, value, sourceStyle.getPropertyPriority(name));
-        });
-    }
-}
-function cloneInputValue(nativeNode, clonedNode) {
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLTextAreaElement)) {
-        clonedNode.innerHTML = nativeNode.value;
-    }
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLInputElement)) {
-        clonedNode.setAttribute('value', nativeNode.value);
-    }
-}
-function cloneSelectValue(nativeNode, clonedNode) {
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(nativeNode, HTMLSelectElement)) {
-        const clonedSelect = clonedNode;
-        const selectedOption = Array.from(clonedSelect.children).find((child) => nativeNode.value === child.getAttribute('value'));
-        if (selectedOption) {
-            selectedOption.setAttribute('selected', '');
-        }
-    }
-}
-function decorate(nativeNode, clonedNode) {
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(clonedNode, Element)) {
-        cloneCSSStyle(nativeNode, clonedNode);
-        (0,_clone_pseudos__WEBPACK_IMPORTED_MODULE_0__.clonePseudoElements)(nativeNode, clonedNode);
-        cloneInputValue(nativeNode, clonedNode);
-        cloneSelectValue(nativeNode, clonedNode);
-    }
-    return clonedNode;
-}
-async function ensureSVGSymbols(clone, options) {
-    const uses = clone.querySelectorAll ? clone.querySelectorAll('use') : [];
-    if (uses.length === 0) {
-        return clone;
-    }
-    const processedDefs = {};
-    for (let i = 0; i < uses.length; i++) {
-        const use = uses[i];
-        const id = use.getAttribute('xlink:href');
-        if (id) {
-            const exist = clone.querySelector(id);
-            const definition = document.querySelector(id);
-            if (!exist && definition && !processedDefs[id]) {
-                // eslint-disable-next-line no-await-in-loop
-                processedDefs[id] = (await cloneNode(definition, options, true));
-            }
-        }
-    }
-    const nodes = Object.values(processedDefs);
-    if (nodes.length) {
-        const ns = 'http://www.w3.org/1999/xhtml';
-        const svg = document.createElementNS(ns, 'svg');
-        svg.setAttribute('xmlns', ns);
-        svg.style.position = 'absolute';
-        svg.style.width = '0';
-        svg.style.height = '0';
-        svg.style.overflow = 'hidden';
-        svg.style.display = 'none';
-        const defs = document.createElementNS(ns, 'defs');
-        svg.appendChild(defs);
-        for (let i = 0; i < nodes.length; i++) {
-            defs.appendChild(nodes[i]);
-        }
-        clone.appendChild(svg);
-    }
-    return clone;
-}
-async function cloneNode(node, options, isRoot) {
-    if (!isRoot && options.filter && !options.filter(node)) {
-        return null;
-    }
-    return Promise.resolve(node)
-        .then((clonedNode) => cloneSingleNode(clonedNode, options))
-        .then((clonedNode) => cloneChildren(node, clonedNode, options))
-        .then((clonedNode) => decorate(node, clonedNode))
-        .then((clonedNode) => ensureSVGSymbols(clonedNode, options));
-}
-//# sourceMappingURL=clone-node.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/clone-pseudos.js":
-/*!********************************************************!*\
-  !*** ./node_modules/html-to-image/es/clone-pseudos.js ***!
-  \********************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   clonePseudoElements: () => (/* binding */ clonePseudoElements)
-/* harmony export */ });
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-
-function formatCSSText(style) {
-    const content = style.getPropertyValue('content');
-    return `${style.cssText} content: '${content.replace(/'|"/g, '')}';`;
-}
-function formatCSSProperties(style) {
-    return (0,_util__WEBPACK_IMPORTED_MODULE_0__.toArray)(style)
-        .map((name) => {
-        const value = style.getPropertyValue(name);
-        const priority = style.getPropertyPriority(name);
-        return `${name}: ${value}${priority ? ' !important' : ''};`;
-    })
-        .join(' ');
-}
-function getPseudoElementStyle(className, pseudo, style) {
-    const selector = `.${className}:${pseudo}`;
-    const cssText = style.cssText
-        ? formatCSSText(style)
-        : formatCSSProperties(style);
-    return document.createTextNode(`${selector}{${cssText}}`);
-}
-function clonePseudoElement(nativeNode, clonedNode, pseudo) {
-    const style = window.getComputedStyle(nativeNode, pseudo);
-    const content = style.getPropertyValue('content');
-    if (content === '' || content === 'none') {
-        return;
-    }
-    const className = (0,_util__WEBPACK_IMPORTED_MODULE_0__.uuid)();
-    try {
-        clonedNode.className = `${clonedNode.className} ${className}`;
-    }
-    catch (err) {
-        return;
-    }
-    const styleElement = document.createElement('style');
-    styleElement.appendChild(getPseudoElementStyle(className, pseudo, style));
-    clonedNode.appendChild(styleElement);
-}
-function clonePseudoElements(nativeNode, clonedNode) {
-    clonePseudoElement(nativeNode, clonedNode, ':before');
-    clonePseudoElement(nativeNode, clonedNode, ':after');
-}
-//# sourceMappingURL=clone-pseudos.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/dataurl.js":
-/*!**************************************************!*\
-  !*** ./node_modules/html-to-image/es/dataurl.js ***!
-  \**************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   fetchAsDataURL: () => (/* binding */ fetchAsDataURL),
-/* harmony export */   isDataUrl: () => (/* binding */ isDataUrl),
-/* harmony export */   makeDataUrl: () => (/* binding */ makeDataUrl),
-/* harmony export */   resourceToDataURL: () => (/* binding */ resourceToDataURL)
-/* harmony export */ });
-function getContentFromDataUrl(dataURL) {
-    return dataURL.split(/,/)[1];
-}
-function isDataUrl(url) {
-    return url.search(/^(data:)/) !== -1;
-}
-function makeDataUrl(content, mimeType) {
-    return `data:${mimeType};base64,${content}`;
-}
-async function fetchAsDataURL(url, init, process) {
-    const res = await fetch(url, init);
-    if (res.status === 404) {
-        throw new Error(`Resource "${res.url}" not found`);
-    }
-    const blob = await res.blob();
-    return new Promise((resolve, reject) => {
-        const reader = new FileReader();
-        reader.onerror = reject;
-        reader.onloadend = () => {
-            try {
-                resolve(process({ res, result: reader.result }));
-            }
-            catch (error) {
-                reject(error);
-            }
-        };
-        reader.readAsDataURL(blob);
-    });
-}
-const cache = {};
-function getCacheKey(url, contentType, includeQueryParams) {
-    let key = url.replace(/\?.*/, '');
-    if (includeQueryParams) {
-        key = url;
-    }
-    // font resource
-    if (/ttf|otf|eot|woff2?/i.test(key)) {
-        key = key.replace(/.*\//, '');
-    }
-    return contentType ? `[${contentType}]${key}` : key;
-}
-async function resourceToDataURL(resourceUrl, contentType, options) {
-    const cacheKey = getCacheKey(resourceUrl, contentType, options.includeQueryParams);
-    if (cache[cacheKey] != null) {
-        return cache[cacheKey];
-    }
-    // ref: https://developer.mozilla.org/en/docs/Web/API/XMLHttpRequest/Using_XMLHttpRequest#Bypassing_the_cache
-    if (options.cacheBust) {
-        // eslint-disable-next-line no-param-reassign
-        resourceUrl += (/\?/.test(resourceUrl) ? '&' : '?') + new Date().getTime();
-    }
-    let dataURL;
-    try {
-        const content = await fetchAsDataURL(resourceUrl, options.fetchRequestInit, ({ res, result }) => {
-            if (!contentType) {
-                // eslint-disable-next-line no-param-reassign
-                contentType = res.headers.get('Content-Type') || '';
-            }
-            return getContentFromDataUrl(result);
-        });
-        dataURL = makeDataUrl(content, contentType);
-    }
-    catch (error) {
-        dataURL = options.imagePlaceholder || '';
-        let msg = `Failed to fetch resource: ${resourceUrl}`;
-        if (error) {
-            msg = typeof error === 'string' ? error : error.message;
-        }
-        if (msg) {
-            console.warn(msg);
-        }
-    }
-    cache[cacheKey] = dataURL;
-    return dataURL;
-}
-//# sourceMappingURL=dataurl.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/embed-images.js":
-/*!*******************************************************!*\
-  !*** ./node_modules/html-to-image/es/embed-images.js ***!
-  \*******************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   embedImages: () => (/* binding */ embedImages)
-/* harmony export */ });
-/* harmony import */ var _embed_resources__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./embed-resources */ "./node_modules/html-to-image/es/embed-resources.js");
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-/* harmony import */ var _dataurl__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./dataurl */ "./node_modules/html-to-image/es/dataurl.js");
-/* harmony import */ var _mimes__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./mimes */ "./node_modules/html-to-image/es/mimes.js");
-
-
-
-
-async function embedProp(propName, node, options) {
-    var _a;
-    const propValue = (_a = node.style) === null || _a === void 0 ? void 0 : _a.getPropertyValue(propName);
-    if (propValue) {
-        const cssString = await (0,_embed_resources__WEBPACK_IMPORTED_MODULE_0__.embedResources)(propValue, null, options);
-        node.style.setProperty(propName, cssString, node.style.getPropertyPriority(propName));
-        return true;
-    }
-    return false;
-}
-async function embedBackground(clonedNode, options) {
-    if (!(await embedProp('background', clonedNode, options))) {
-        await embedProp('background-image', clonedNode, options);
-    }
-    if (!(await embedProp('mask', clonedNode, options))) {
-        await embedProp('mask-image', clonedNode, options);
-    }
-}
-async function embedImageNode(clonedNode, options) {
-    const isImageElement = (0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(clonedNode, HTMLImageElement);
-    if (!(isImageElement && !(0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.isDataUrl)(clonedNode.src)) &&
-        !((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(clonedNode, SVGImageElement) &&
-            !(0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.isDataUrl)(clonedNode.href.baseVal))) {
-        return;
-    }
-    const url = isImageElement ? clonedNode.src : clonedNode.href.baseVal;
-    const dataURL = await (0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.resourceToDataURL)(url, (0,_mimes__WEBPACK_IMPORTED_MODULE_3__.getMimeType)(url), options);
-    await new Promise((resolve, reject) => {
-        clonedNode.onload = resolve;
-        clonedNode.onerror = reject;
-        const image = clonedNode;
-        if (image.decode) {
-            image.decode = resolve;
-        }
-        if (image.loading === 'lazy') {
-            image.loading = 'eager';
-        }
-        if (isImageElement) {
-            clonedNode.srcset = '';
-            clonedNode.src = dataURL;
-        }
-        else {
-            clonedNode.href.baseVal = dataURL;
-        }
-    });
-}
-async function embedChildren(clonedNode, options) {
-    const children = (0,_util__WEBPACK_IMPORTED_MODULE_1__.toArray)(clonedNode.childNodes);
-    const deferreds = children.map((child) => embedImages(child, options));
-    await Promise.all(deferreds).then(() => clonedNode);
-}
-async function embedImages(clonedNode, options) {
-    if ((0,_util__WEBPACK_IMPORTED_MODULE_1__.isInstanceOfElement)(clonedNode, Element)) {
-        await embedBackground(clonedNode, options);
-        await embedImageNode(clonedNode, options);
-        await embedChildren(clonedNode, options);
-    }
-}
-//# sourceMappingURL=embed-images.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/embed-resources.js":
-/*!**********************************************************!*\
-  !*** ./node_modules/html-to-image/es/embed-resources.js ***!
-  \**********************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   embed: () => (/* binding */ embed),
-/* harmony export */   embedResources: () => (/* binding */ embedResources),
-/* harmony export */   parseURLs: () => (/* binding */ parseURLs),
-/* harmony export */   shouldEmbed: () => (/* binding */ shouldEmbed)
-/* harmony export */ });
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-/* harmony import */ var _mimes__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./mimes */ "./node_modules/html-to-image/es/mimes.js");
-/* harmony import */ var _dataurl__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./dataurl */ "./node_modules/html-to-image/es/dataurl.js");
-
-
-
-const URL_REGEX = /url\((['"]?)([^'"]+?)\1\)/g;
-const URL_WITH_FORMAT_REGEX = /url\([^)]+\)\s*format\((["']?)([^"']+)\1\)/g;
-const FONT_SRC_REGEX = /src:\s*(?:url\([^)]+\)\s*format\([^)]+\)[,;]\s*)+/g;
-function toRegex(url) {
-    // eslint-disable-next-line no-useless-escape
-    const escaped = url.replace(/([.*+?^${}()|\[\]\/\\])/g, '\\$1');
-    return new RegExp(`(url\\(['"]?)(${escaped})(['"]?\\))`, 'g');
-}
-function parseURLs(cssText) {
-    const urls = [];
-    cssText.replace(URL_REGEX, (raw, quotation, url) => {
-        urls.push(url);
-        return raw;
-    });
-    return urls.filter((url) => !(0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.isDataUrl)(url));
-}
-async function embed(cssText, resourceURL, baseURL, options, getContentFromUrl) {
-    try {
-        const resolvedURL = baseURL ? (0,_util__WEBPACK_IMPORTED_MODULE_0__.resolveUrl)(resourceURL, baseURL) : resourceURL;
-        const contentType = (0,_mimes__WEBPACK_IMPORTED_MODULE_1__.getMimeType)(resourceURL);
-        let dataURL;
-        if (getContentFromUrl) {
-            const content = await getContentFromUrl(resolvedURL);
-            dataURL = (0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.makeDataUrl)(content, contentType);
-        }
-        else {
-            dataURL = await (0,_dataurl__WEBPACK_IMPORTED_MODULE_2__.resourceToDataURL)(resolvedURL, contentType, options);
-        }
-        return cssText.replace(toRegex(resourceURL), `$1${dataURL}$3`);
-    }
-    catch (error) {
-        // pass
-    }
-    return cssText;
-}
-function filterPreferredFontFormat(str, { preferredFontFormat }) {
-    return !preferredFontFormat
-        ? str
-        : str.replace(FONT_SRC_REGEX, (match) => {
-            // eslint-disable-next-line no-constant-condition
-            while (true) {
-                const [src, , format] = URL_WITH_FORMAT_REGEX.exec(match) || [];
-                if (!format) {
-                    return '';
-                }
-                if (format === preferredFontFormat) {
-                    return `src: ${src};`;
-                }
-            }
-        });
-}
-function shouldEmbed(url) {
-    return url.search(URL_REGEX) !== -1;
-}
-async function embedResources(cssText, baseUrl, options) {
-    if (!shouldEmbed(cssText)) {
-        return cssText;
-    }
-    const filteredCSSText = filterPreferredFontFormat(cssText, options);
-    const urls = parseURLs(filteredCSSText);
-    return urls.reduce((deferred, url) => deferred.then((css) => embed(css, url, baseUrl, options)), Promise.resolve(filteredCSSText));
-}
-//# sourceMappingURL=embed-resources.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/embed-webfonts.js":
-/*!*********************************************************!*\
-  !*** ./node_modules/html-to-image/es/embed-webfonts.js ***!
-  \*********************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   embedWebFonts: () => (/* binding */ embedWebFonts),
-/* harmony export */   getWebFontCSS: () => (/* binding */ getWebFontCSS)
-/* harmony export */ });
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-/* harmony import */ var _dataurl__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./dataurl */ "./node_modules/html-to-image/es/dataurl.js");
-/* harmony import */ var _embed_resources__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./embed-resources */ "./node_modules/html-to-image/es/embed-resources.js");
-
-
-
-const cssFetchCache = {};
-async function fetchCSS(url) {
-    let cache = cssFetchCache[url];
-    if (cache != null) {
-        return cache;
-    }
-    const res = await fetch(url);
-    const cssText = await res.text();
-    cache = { url, cssText };
-    cssFetchCache[url] = cache;
-    return cache;
-}
-async function embedFonts(data, options) {
-    let cssText = data.cssText;
-    const regexUrl = /url\(["']?([^"')]+)["']?\)/g;
-    const fontLocs = cssText.match(/url\([^)]+\)/g) || [];
-    const loadFonts = fontLocs.map(async (loc) => {
-        let url = loc.replace(regexUrl, '$1');
-        if (!url.startsWith('https://')) {
-            url = new URL(url, data.url).href;
-        }
-        return (0,_dataurl__WEBPACK_IMPORTED_MODULE_1__.fetchAsDataURL)(url, options.fetchRequestInit, ({ result }) => {
-            cssText = cssText.replace(loc, `url(${result})`);
-            return [loc, result];
-        });
-    });
-    return Promise.all(loadFonts).then(() => cssText);
-}
-function parseCSS(source) {
-    if (source == null) {
-        return [];
-    }
-    const result = [];
-    const commentsRegex = /(\/\*[\s\S]*?\*\/)/gi;
-    // strip out comments
-    let cssText = source.replace(commentsRegex, '');
-    // eslint-disable-next-line prefer-regex-literals
-    const keyframesRegex = new RegExp('((@.*?keyframes [\\s\\S]*?){([\\s\\S]*?}\\s*?)})', 'gi');
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-        const matches = keyframesRegex.exec(cssText);
-        if (matches === null) {
-            break;
-        }
-        result.push(matches[0]);
-    }
-    cssText = cssText.replace(keyframesRegex, '');
-    const importRegex = /@import[\s\S]*?url\([^)]*\)[\s\S]*?;/gi;
-    // to match css & media queries together
-    const combinedCSSRegex = '((\\s*?(?:\\/\\*[\\s\\S]*?\\*\\/)?\\s*?@media[\\s\\S]' +
-        '*?){([\\s\\S]*?)}\\s*?})|(([\\s\\S]*?){([\\s\\S]*?)})';
-    // unified regex
-    const unifiedRegex = new RegExp(combinedCSSRegex, 'gi');
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-        let matches = importRegex.exec(cssText);
-        if (matches === null) {
-            matches = unifiedRegex.exec(cssText);
-            if (matches === null) {
-                break;
-            }
-            else {
-                importRegex.lastIndex = unifiedRegex.lastIndex;
-            }
-        }
-        else {
-            unifiedRegex.lastIndex = importRegex.lastIndex;
-        }
-        result.push(matches[0]);
-    }
-    return result;
-}
-async function getCSSRules(styleSheets, options) {
-    const ret = [];
-    const deferreds = [];
-    // First loop inlines imports
-    styleSheets.forEach((sheet) => {
-        if ('cssRules' in sheet) {
-            try {
-                (0,_util__WEBPACK_IMPORTED_MODULE_0__.toArray)(sheet.cssRules || []).forEach((item, index) => {
-                    if (item.type === CSSRule.IMPORT_RULE) {
-                        let importIndex = index + 1;
-                        const url = item.href;
-                        const deferred = fetchCSS(url)
-                            .then((metadata) => embedFonts(metadata, options))
-                            .then((cssText) => parseCSS(cssText).forEach((rule) => {
-                            try {
-                                sheet.insertRule(rule, rule.startsWith('@import')
-                                    ? (importIndex += 1)
-                                    : sheet.cssRules.length);
-                            }
-                            catch (error) {
-                                console.error('Error inserting rule from remote css', {
-                                    rule,
-                                    error,
-                                });
-                            }
-                        }))
-                            .catch((e) => {
-                            console.error('Error loading remote css', e.toString());
-                        });
-                        deferreds.push(deferred);
-                    }
-                });
-            }
-            catch (e) {
-                const inline = styleSheets.find((a) => a.href == null) || document.styleSheets[0];
-                if (sheet.href != null) {
-                    deferreds.push(fetchCSS(sheet.href)
-                        .then((metadata) => embedFonts(metadata, options))
-                        .then((cssText) => parseCSS(cssText).forEach((rule) => {
-                        inline.insertRule(rule, sheet.cssRules.length);
-                    }))
-                        .catch((err) => {
-                        console.error('Error loading remote stylesheet', err);
-                    }));
-                }
-                console.error('Error inlining remote css file', e);
-            }
-        }
-    });
-    return Promise.all(deferreds).then(() => {
-        // Second loop parses rules
-        styleSheets.forEach((sheet) => {
-            if ('cssRules' in sheet) {
-                try {
-                    (0,_util__WEBPACK_IMPORTED_MODULE_0__.toArray)(sheet.cssRules || []).forEach((item) => {
-                        ret.push(item);
-                    });
-                }
-                catch (e) {
-                    console.error(`Error while reading CSS rules from ${sheet.href}`, e);
-                }
-            }
-        });
-        return ret;
-    });
-}
-function getWebFontRules(cssRules) {
-    return cssRules
-        .filter((rule) => rule.type === CSSRule.FONT_FACE_RULE)
-        .filter((rule) => (0,_embed_resources__WEBPACK_IMPORTED_MODULE_2__.shouldEmbed)(rule.style.getPropertyValue('src')));
-}
-async function parseWebFontRules(node, options) {
-    if (node.ownerDocument == null) {
-        throw new Error('Provided element is not within a Document');
-    }
-    const styleSheets = (0,_util__WEBPACK_IMPORTED_MODULE_0__.toArray)(node.ownerDocument.styleSheets);
-    const cssRules = await getCSSRules(styleSheets, options);
-    return getWebFontRules(cssRules);
-}
-async function getWebFontCSS(node, options) {
-    const rules = await parseWebFontRules(node, options);
-    const cssTexts = await Promise.all(rules.map((rule) => {
-        const baseUrl = rule.parentStyleSheet ? rule.parentStyleSheet.href : null;
-        return (0,_embed_resources__WEBPACK_IMPORTED_MODULE_2__.embedResources)(rule.cssText, baseUrl, options);
-    }));
-    return cssTexts.join('\n');
-}
-async function embedWebFonts(clonedNode, options) {
-    const cssText = options.fontEmbedCSS != null
-        ? options.fontEmbedCSS
-        : options.skipFonts
-            ? null
-            : await getWebFontCSS(clonedNode, options);
-    if (cssText) {
-        const styleNode = document.createElement('style');
-        const sytleContent = document.createTextNode(cssText);
-        styleNode.appendChild(sytleContent);
-        if (clonedNode.firstChild) {
-            clonedNode.insertBefore(styleNode, clonedNode.firstChild);
-        }
-        else {
-            clonedNode.appendChild(styleNode);
-        }
-    }
-}
-//# sourceMappingURL=embed-webfonts.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/index.js":
-/*!************************************************!*\
-  !*** ./node_modules/html-to-image/es/index.js ***!
-  \************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   getFontEmbedCSS: () => (/* binding */ getFontEmbedCSS),
-/* harmony export */   toBlob: () => (/* binding */ toBlob),
-/* harmony export */   toCanvas: () => (/* binding */ toCanvas),
-/* harmony export */   toJpeg: () => (/* binding */ toJpeg),
-/* harmony export */   toPixelData: () => (/* binding */ toPixelData),
-/* harmony export */   toPng: () => (/* binding */ toPng),
-/* harmony export */   toSvg: () => (/* binding */ toSvg)
-/* harmony export */ });
-/* harmony import */ var _clone_node__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./clone-node */ "./node_modules/html-to-image/es/clone-node.js");
-/* harmony import */ var _embed_images__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./embed-images */ "./node_modules/html-to-image/es/embed-images.js");
-/* harmony import */ var _apply_style__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./apply-style */ "./node_modules/html-to-image/es/apply-style.js");
-/* harmony import */ var _embed_webfonts__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./embed-webfonts */ "./node_modules/html-to-image/es/embed-webfonts.js");
-/* harmony import */ var _util__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./util */ "./node_modules/html-to-image/es/util.js");
-
-
-
-
-
-async function toSvg(node, options = {}) {
-    const { width, height } = (0,_util__WEBPACK_IMPORTED_MODULE_4__.getImageSize)(node, options);
-    const clonedNode = (await (0,_clone_node__WEBPACK_IMPORTED_MODULE_0__.cloneNode)(node, options, true));
-    await (0,_embed_webfonts__WEBPACK_IMPORTED_MODULE_3__.embedWebFonts)(clonedNode, options);
-    await (0,_embed_images__WEBPACK_IMPORTED_MODULE_1__.embedImages)(clonedNode, options);
-    (0,_apply_style__WEBPACK_IMPORTED_MODULE_2__.applyStyle)(clonedNode, options);
-    const datauri = await (0,_util__WEBPACK_IMPORTED_MODULE_4__.nodeToDataURL)(clonedNode, width, height);
-    return datauri;
-}
-async function toCanvas(node, options = {}) {
-    const { width, height } = (0,_util__WEBPACK_IMPORTED_MODULE_4__.getImageSize)(node, options);
-    const svg = await toSvg(node, options);
-    const img = await (0,_util__WEBPACK_IMPORTED_MODULE_4__.createImage)(svg);
-    const canvas = document.createElement('canvas');
-    const context = canvas.getContext('2d');
-    const ratio = options.pixelRatio || (0,_util__WEBPACK_IMPORTED_MODULE_4__.getPixelRatio)();
-    const canvasWidth = options.canvasWidth || width;
-    const canvasHeight = options.canvasHeight || height;
-    canvas.width = canvasWidth * ratio;
-    canvas.height = canvasHeight * ratio;
-    if (!options.skipAutoScale) {
-        (0,_util__WEBPACK_IMPORTED_MODULE_4__.checkCanvasDimensions)(canvas);
-    }
-    canvas.style.width = `${canvasWidth}`;
-    canvas.style.height = `${canvasHeight}`;
-    if (options.backgroundColor) {
-        context.fillStyle = options.backgroundColor;
-        context.fillRect(0, 0, canvas.width, canvas.height);
-    }
-    context.drawImage(img, 0, 0, canvas.width, canvas.height);
-    return canvas;
-}
-async function toPixelData(node, options = {}) {
-    const { width, height } = (0,_util__WEBPACK_IMPORTED_MODULE_4__.getImageSize)(node, options);
-    const canvas = await toCanvas(node, options);
-    const ctx = canvas.getContext('2d');
-    return ctx.getImageData(0, 0, width, height).data;
-}
-async function toPng(node, options = {}) {
-    const canvas = await toCanvas(node, options);
-    return canvas.toDataURL();
-}
-async function toJpeg(node, options = {}) {
-    const canvas = await toCanvas(node, options);
-    return canvas.toDataURL('image/jpeg', options.quality || 1);
-}
-async function toBlob(node, options = {}) {
-    const canvas = await toCanvas(node, options);
-    const blob = await (0,_util__WEBPACK_IMPORTED_MODULE_4__.canvasToBlob)(canvas);
-    return blob;
-}
-async function getFontEmbedCSS(node, options = {}) {
-    return (0,_embed_webfonts__WEBPACK_IMPORTED_MODULE_3__.getWebFontCSS)(node, options);
-}
-//# sourceMappingURL=index.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/mimes.js":
-/*!************************************************!*\
-  !*** ./node_modules/html-to-image/es/mimes.js ***!
-  \************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   getMimeType: () => (/* binding */ getMimeType)
-/* harmony export */ });
-const WOFF = 'application/font-woff';
-const JPEG = 'image/jpeg';
-const mimes = {
-    woff: WOFF,
-    woff2: WOFF,
-    ttf: 'application/font-truetype',
-    eot: 'application/vnd.ms-fontobject',
-    png: 'image/png',
-    jpg: JPEG,
-    jpeg: JPEG,
-    gif: 'image/gif',
-    tiff: 'image/tiff',
-    svg: 'image/svg+xml',
-    webp: 'image/webp',
-};
-function getExtension(url) {
-    const match = /\.([^./]*?)$/g.exec(url);
-    return match ? match[1] : '';
-}
-function getMimeType(url) {
-    const extension = getExtension(url).toLowerCase();
-    return mimes[extension] || '';
-}
-//# sourceMappingURL=mimes.js.map
-
-/***/ }),
-
-/***/ "./node_modules/html-to-image/es/util.js":
-/*!***********************************************!*\
-  !*** ./node_modules/html-to-image/es/util.js ***!
-  \***********************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   canvasToBlob: () => (/* binding */ canvasToBlob),
-/* harmony export */   checkCanvasDimensions: () => (/* binding */ checkCanvasDimensions),
-/* harmony export */   createImage: () => (/* binding */ createImage),
-/* harmony export */   delay: () => (/* binding */ delay),
-/* harmony export */   getImageSize: () => (/* binding */ getImageSize),
-/* harmony export */   getPixelRatio: () => (/* binding */ getPixelRatio),
-/* harmony export */   isInstanceOfElement: () => (/* binding */ isInstanceOfElement),
-/* harmony export */   nodeToDataURL: () => (/* binding */ nodeToDataURL),
-/* harmony export */   resolveUrl: () => (/* binding */ resolveUrl),
-/* harmony export */   svgToDataURL: () => (/* binding */ svgToDataURL),
-/* harmony export */   toArray: () => (/* binding */ toArray),
-/* harmony export */   uuid: () => (/* binding */ uuid)
-/* harmony export */ });
-function resolveUrl(url, baseUrl) {
-    // url is absolute already
-    if (url.match(/^[a-z]+:\/\//i)) {
-        return url;
-    }
-    // url is absolute already, without protocol
-    if (url.match(/^\/\//)) {
-        return window.location.protocol + url;
-    }
-    // dataURI, mailto:, tel:, etc.
-    if (url.match(/^[a-z]+:/i)) {
-        return url;
-    }
-    const doc = document.implementation.createHTMLDocument();
-    const base = doc.createElement('base');
-    const a = doc.createElement('a');
-    doc.head.appendChild(base);
-    doc.body.appendChild(a);
-    if (baseUrl) {
-        base.href = baseUrl;
-    }
-    a.href = url;
-    return a.href;
-}
-const uuid = (() => {
-    // generate uuid for className of pseudo elements.
-    // We should not use GUIDs, otherwise pseudo elements sometimes cannot be captured.
-    let counter = 0;
-    // ref: http://stackoverflow.com/a/6248722/2519373
-    const random = () => 
-    // eslint-disable-next-line no-bitwise
-    `0000${((Math.random() * 36 ** 4) << 0).toString(36)}`.slice(-4);
-    return () => {
-        counter += 1;
-        return `u${random()}${counter}`;
-    };
-})();
-function delay(ms) {
-    return (args) => new Promise((resolve) => {
-        setTimeout(() => resolve(args), ms);
-    });
-}
-function toArray(arrayLike) {
-    const arr = [];
-    for (let i = 0, l = arrayLike.length; i < l; i++) {
-        arr.push(arrayLike[i]);
-    }
-    return arr;
-}
-function px(node, styleProperty) {
-    const win = node.ownerDocument.defaultView || window;
-    const val = win.getComputedStyle(node).getPropertyValue(styleProperty);
-    return val ? parseFloat(val.replace('px', '')) : 0;
-}
-function getNodeWidth(node) {
-    const leftBorder = px(node, 'border-left-width');
-    const rightBorder = px(node, 'border-right-width');
-    return node.clientWidth + leftBorder + rightBorder;
-}
-function getNodeHeight(node) {
-    const topBorder = px(node, 'border-top-width');
-    const bottomBorder = px(node, 'border-bottom-width');
-    return node.clientHeight + topBorder + bottomBorder;
-}
-function getImageSize(targetNode, options = {}) {
-    const width = options.width || getNodeWidth(targetNode);
-    const height = options.height || getNodeHeight(targetNode);
-    return { width, height };
-}
-function getPixelRatio() {
-    let ratio;
-    let FINAL_PROCESS;
-    try {
-        FINAL_PROCESS = process;
-    }
-    catch (e) {
-        // pass
-    }
-    const val = FINAL_PROCESS && FINAL_PROCESS.env
-        ? FINAL_PROCESS.env.devicePixelRatio
-        : null;
-    if (val) {
-        ratio = parseInt(val, 10);
-        if (Number.isNaN(ratio)) {
-            ratio = 1;
-        }
-    }
-    return ratio || window.devicePixelRatio || 1;
-}
-// @see https://developer.mozilla.org/en-US/docs/Web/HTML/Element/canvas#maximum_canvas_size
-const canvasDimensionLimit = 16384;
-function checkCanvasDimensions(canvas) {
-    if (canvas.width > canvasDimensionLimit ||
-        canvas.height > canvasDimensionLimit) {
-        if (canvas.width > canvasDimensionLimit &&
-            canvas.height > canvasDimensionLimit) {
-            if (canvas.width > canvas.height) {
-                canvas.height *= canvasDimensionLimit / canvas.width;
-                canvas.width = canvasDimensionLimit;
-            }
-            else {
-                canvas.width *= canvasDimensionLimit / canvas.height;
-                canvas.height = canvasDimensionLimit;
-            }
-        }
-        else if (canvas.width > canvasDimensionLimit) {
-            canvas.height *= canvasDimensionLimit / canvas.width;
-            canvas.width = canvasDimensionLimit;
-        }
-        else {
-            canvas.width *= canvasDimensionLimit / canvas.height;
-            canvas.height = canvasDimensionLimit;
-        }
-    }
-}
-function canvasToBlob(canvas, options = {}) {
-    if (canvas.toBlob) {
-        return new Promise((resolve) => {
-            canvas.toBlob(resolve, options.type ? options.type : 'image/png', options.quality ? options.quality : 1);
-        });
-    }
-    return new Promise((resolve) => {
-        const binaryString = window.atob(canvas
-            .toDataURL(options.type ? options.type : undefined, options.quality ? options.quality : undefined)
-            .split(',')[1]);
-        const len = binaryString.length;
-        const binaryArray = new Uint8Array(len);
-        for (let i = 0; i < len; i += 1) {
-            binaryArray[i] = binaryString.charCodeAt(i);
-        }
-        resolve(new Blob([binaryArray], {
-            type: options.type ? options.type : 'image/png',
-        }));
-    });
-}
-function createImage(url) {
-    return new Promise((resolve, reject) => {
-        const img = new Image();
-        img.decode = () => resolve(img);
-        img.onload = () => resolve(img);
-        img.onerror = reject;
-        img.crossOrigin = 'anonymous';
-        img.decoding = 'async';
-        img.src = url;
-    });
-}
-async function svgToDataURL(svg) {
-    return Promise.resolve()
-        .then(() => new XMLSerializer().serializeToString(svg))
-        .then(encodeURIComponent)
-        .then((html) => `data:image/svg+xml;charset=utf-8,${html}`);
-}
-async function nodeToDataURL(node, width, height) {
-    const xmlns = 'http://www.w3.org/2000/svg';
-    const svg = document.createElementNS(xmlns, 'svg');
-    const foreignObject = document.createElementNS(xmlns, 'foreignObject');
-    svg.setAttribute('width', `${width}`);
-    svg.setAttribute('height', `${height}`);
-    svg.setAttribute('viewBox', `0 0 ${width} ${height}`);
-    foreignObject.setAttribute('width', '100%');
-    foreignObject.setAttribute('height', '100%');
-    foreignObject.setAttribute('x', '0');
-    foreignObject.setAttribute('y', '0');
-    foreignObject.setAttribute('externalResourcesRequired', 'true');
-    svg.appendChild(foreignObject);
-    foreignObject.appendChild(node);
-    return svgToDataURL(svg);
-}
-const isInstanceOfElement = (node, instance) => {
-    if (node instanceof instance)
-        return true;
-    const nodePrototype = Object.getPrototypeOf(node);
-    if (nodePrototype === null)
-        return false;
-    return (nodePrototype.constructor.name === instance.name ||
-        isInstanceOfElement(nodePrototype, instance));
-};
-//# sourceMappingURL=util.js.map
 
 /***/ }),
