+++ assets/jsx/workflow-editor/components/workflow-store/selectors.jsx
@@ -1,6 +1,3 @@
-// import { createRegistrySelector } from '@wordpress/data';
-import { toPng } from 'html-to-image';
-
 export const getPostType = (state) => {
     return state.postType;
 };
@@ -145,33 +142,31 @@ export const getDataTypeByName = (state, name) => {
     return state.dataTypes.find(dataType => dataType.name === name);
 }
 
-export function takeScreenshot() {
-    return new Promise((resolve, reject) => {
-        try {
-            toPng(document.querySelector('.react-flow'), {
-                filter: (node) => {
-                    // We don't want to add the minimap and the controls to the image
-                    if (
-                        node?.classList?.contains('react-flow__minimap') ||
-                        node?.classList?.contains('react-flow__controls') ||
-                        node?.classList?.contains('pwe-node-edit-button')
-                    ) {
-                        return false;
-                    }
+export async function takeScreenshot() {
+    const { enableWorkflowScreenshot } = futureWorkflowEditor;
 
-                    return true;
-                },
-            })
-                .then((dataUrl) => {
-                    resolve(dataUrl); // Resolve with the image dataUrl
-                })
-                .catch((error) => {
-                    reject(error); // Reject with the error
-                });
-        } catch (error) {
-            reject(error); // Reject with the error
-        }
-    });
+    if (!enableWorkflowScreenshot) {
+        return null;
+    }
+
+    try {
+        const { toPng } = await import('html-to-image');
+        return await toPng(document.querySelector('.react-flow'), {
+            filter: (node) => {
+                if (
+                    node?.classList?.contains('react-flow__minimap') ||
+                    node?.classList?.contains('react-flow__controls') ||
+                    node?.classList?.contains('pwe-node-edit-button')
+                ) {
+                    return false;
+                }
+                return true;
+            },
+        });
+    } catch (error) {
+        console.error('Error loading html-to-image:', error);
+        return null;
+    }
 }
 
 export function getGlobalVariables(state) {
