+++ languages/post-expirator-future_backup_panel.pot
@@ -7,18 +7,18 @@ msgstr ""
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"POT-Creation-Date: 2024-12-05T13:02:37+00:00\n"
+"POT-Creation-Date: 2024-12-05T19:40:46+00:00\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "X-Generator: WP-CLI 2.11.0\n"
 "X-Domain: post-expirator\n"
 
 #: components/backup-panel.jsx:13
-#: components/tabs-content/export.jsx:201
+#: components/tabs-content/export.jsx:205
 msgid "Export"
 msgstr ""
 
 #: components/backup-panel.jsx:17
-#: components/tabs-content/import.jsx:144
+#: components/tabs-content/import.jsx:148
 msgid "Import"
 msgstr ""
 
@@ -30,102 +30,102 @@ msgstr ""
 msgid "Unselect all"
 msgstr ""
 
-#: components/tabs-content/export.jsx:23
+#: components/tabs-content/export.jsx:25
 msgid "Post Types"
 msgstr ""
 
-#: components/tabs-content/export.jsx:27
+#: components/tabs-content/export.jsx:29
 msgid "General"
 msgstr ""
 
-#: components/tabs-content/export.jsx:31
+#: components/tabs-content/export.jsx:33
 msgid "Notifications"
 msgstr ""
 
-#: components/tabs-content/export.jsx:35
+#: components/tabs-content/export.jsx:37
 msgid "Display"
 msgstr ""
 
-#: components/tabs-content/export.jsx:39
+#: components/tabs-content/export.jsx:41
 msgid "Advanced"
 msgstr ""
 
-#: components/tabs-content/export.jsx:102
+#: components/tabs-content/export.jsx:104
 msgid "Settings exported successfully."
 msgstr ""
 
-#: components/tabs-content/export.jsx:108
+#: components/tabs-content/export.jsx:110
 msgid "Download"
 msgstr ""
 
-#: components/tabs-content/export.jsx:125
+#: components/tabs-content/export.jsx:127
 msgid "Failed to export settings."
 msgstr ""
 
-#: components/tabs-content/export.jsx:158
+#: components/tabs-content/export.jsx:160
 msgid "Export Settings"
 msgstr ""
 
-#: components/tabs-content/export.jsx:159
+#: components/tabs-content/export.jsx:161
 msgid "Export the plugin settings and workflows to a .json file. This file can be imported later to restore the data or migrate to another site."
 msgstr ""
 
-#: components/tabs-content/export.jsx:164
+#: components/tabs-content/export.jsx:166
 msgid "Action Workflows"
 msgstr ""
 
-#: components/tabs-content/export.jsx:174
+#: components/tabs-content/export.jsx:177
 msgid "Include screenshots"
 msgstr ""
 
-#: components/tabs-content/export.jsx:186
+#: components/tabs-content/export.jsx:190
 msgid "Action Settings"
 msgstr ""
 
-#: components/tabs-content/export.jsx:201
+#: components/tabs-content/export.jsx:205
 msgid "Exporting..."
 msgstr ""
 
-#: components/tabs-content/export.jsx:205
+#: components/tabs-content/export.jsx:209
 msgid "Cancel"
 msgstr ""
 
-#: components/tabs-content/import.jsx:47
+#: components/tabs-content/import.jsx:48
 msgid "Settings imported successfully."
 msgstr ""
 
-#: components/tabs-content/import.jsx:61
+#: components/tabs-content/import.jsx:65
 msgid "Failed to import settings."
 msgstr ""
 
-#: components/tabs-content/import.jsx:78
+#: components/tabs-content/import.jsx:83
 msgid "Invalid file type. Please upload a .json file."
 msgstr ""
 
-#: components/tabs-content/import.jsx:90
+#: components/tabs-content/import.jsx:94
 msgid "Import Settings"
 msgstr ""
 
-#: components/tabs-content/import.jsx:91
+#: components/tabs-content/import.jsx:95
 msgid "Import the plugin settings or workflows from a .json file."
 msgstr ""
 
-#: components/tabs-content/import.jsx:118
+#: components/tabs-content/import.jsx:122
 msgid "Drop your .json file here"
 msgstr ""
 
-#: components/tabs-content/import.jsx:119
+#: components/tabs-content/import.jsx:123
 msgid "or"
 msgstr ""
 
-#: components/tabs-content/import.jsx:129
+#: components/tabs-content/import.jsx:133
 msgid "Select a .json file"
 msgstr ""
 
-#: components/tabs-content/import.jsx:134
+#: components/tabs-content/import.jsx:138
 msgid "Selected file"
 msgstr ""
 
-#: components/tabs-content/import.jsx:135
+#: components/tabs-content/import.jsx:139
 msgid "File size"
 msgstr ""
