+++ languages/post-expirator-future_workflow_editor_script.pot
@@ -7,7 +7,7 @@ msgstr ""
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"POT-Creation-Date: 2024-12-05T13:02:37+00:00\n"
+"POT-Creation-Date: 2024-12-05T19:40:46+00:00\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "X-Generator: WP-CLI 2.11.0\n"
 "X-Domain: post-expirator\n"
@@ -272,26 +272,26 @@ msgstr ""
 msgid "Currently this step is being skipped. Upgrade to Pro to unlock this feature."
 msgstr ""
 
-#: components/layout/header.jsx:133
+#: components/layout/header.jsx:129
 msgid "Add"
 msgstr ""
 
-#: components/layout/header.jsx:133
+#: components/layout/header.jsx:129
 msgid "Close"
 msgstr ""
 
 #. translators: button label text should, if possible, be under 16 characters.
-#: components/layout/header.jsx:144
-#: components/layout/header.jsx:149
+#: components/layout/header.jsx:140
+#: components/layout/header.jsx:145
 msgid "Auto Layout"
 msgstr ""
 
 #. translators: button label text should, if possible, be under 16 characters.
-#: components/layout/header.jsx:161
+#: components/layout/header.jsx:157
 msgid "Upgrade to Pro and unlock more features"
 msgstr ""
 
-#: components/layout/header.jsx:164
+#: components/layout/header.jsx:160
 msgid "Upgrade to Pro"
 msgstr ""
 
@@ -741,45 +741,45 @@ msgid "Move to trash"
 msgstr ""
 
 #. translators: button label text should, if possible, be under 16 characters.
-#: components/workflow-publish-button/index.jsx:82
+#: components/workflow-publish-button/index.jsx:84
 msgid "Publish"
 msgstr ""
 
-#: components/workflow-publish-button/index.jsx:82
+#: components/workflow-publish-button/index.jsx:84
 msgid "Update"
 msgstr ""
 
-#: components/workflow-publish-button/index.jsx:93
+#: components/workflow-publish-button/index.jsx:95
 msgid "Autosaving..."
 msgstr ""
 
-#: components/workflow-publish-button/index.jsx:93
+#: components/workflow-publish-button/index.jsx:95
 msgid "Updating..."
 msgstr ""
 
-#: components/workflow-publish-button/index.jsx:95
+#: components/workflow-publish-button/index.jsx:97
 msgid "Updated"
 msgstr ""
 
 #. translators: button label text should, if possible, be under 16 characters.
-#: components/workflow-save-draft-button/index.jsx:85
+#: components/workflow-save-draft-button/index.jsx:87
 msgid "Save draft"
 msgstr ""
 
 #. translators: button label text should, if possible, be under 16 characters.
-#: components/workflow-save-draft-button/index.jsx:88
+#: components/workflow-save-draft-button/index.jsx:90
 msgid "Save"
 msgstr ""
 
-#: components/workflow-save-draft-button/index.jsx:97
+#: components/workflow-save-draft-button/index.jsx:99
 msgid "Autosaving"
 msgstr ""
 
-#: components/workflow-save-draft-button/index.jsx:97
+#: components/workflow-save-draft-button/index.jsx:99
 msgid "Saving"
 msgstr ""
 
-#: components/workflow-save-draft-button/index.jsx:99
+#: components/workflow-save-draft-button/index.jsx:101
 msgid "Saved"
 msgstr ""
 
@@ -804,27 +804,27 @@ msgstr ""
 msgid "Workflow saved."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:169
+#: components/workflow-store/actions.jsx:170
 msgid "Workflow published."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:177
+#: components/workflow-store/actions.jsx:178
 msgid "Unable to publish the workflow. Please, try again."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:214
+#: components/workflow-store/actions.jsx:216
 msgid "Workflow switched to draft."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:222
+#: components/workflow-store/actions.jsx:224
 msgid "Unable to switch workflow to draft. Please, try again."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:312
+#: components/workflow-store/actions.jsx:314
 msgid "Workflow deleted. Redirecting..."
 msgstr ""
 
-#: components/workflow-store/actions.jsx:323
+#: components/workflow-store/actions.jsx:325
 msgid "Unable to delete the workflow. Please, try again."
 msgstr ""
 
@@ -840,14 +840,14 @@ msgstr ""
 msgid "Description"
 msgstr ""
 
-#: components/workflow-switch-to-draft-button/index.jsx:35
+#: components/workflow-switch-to-draft-button/index.jsx:37
 msgid "Are you sure you want to unpublish this workflow?"
 msgstr ""
 
-#: components/workflow-switch-to-draft-button/index.jsx:56
+#: components/workflow-switch-to-draft-button/index.jsx:62
 msgid "Draft"
 msgstr ""
 
-#: components/workflow-switch-to-draft-button/index.jsx:56
+#: components/workflow-switch-to-draft-button/index.jsx:62
 msgid "Switch to draft"
 msgstr ""
